Code added to create a figure of 9 largest municipalities including either hospitalizations or positive tests.
To make the figures with the number of hospitalizations more consistent, the figure of the viral load in The Netherlands with the number of hospitalizations is now given by hospitalizations per 100.000 inhabitants.

In df_viralload_human_regions the inhabitants of STPs were double counted for the safety regions, this is now fixed.